metatags, updated dashboard, added settings section and features. some bugs still present with routes and queries.

+++ routes/web.php
@@ -1,6 +1,11 @@
 <?php
 
 use Illuminate\Support\Facades\Route;
+use App\GeneralSetting;
+use App\SocialSetting;
+use App\SeoSetting;
+
+
 
 /*
 |--------------------------------------------------------------------------
@@ -35,19 +40,19 @@ Route::get('/offers/thank-you', 'StaticPagesController@thankYou');
 //Admin 
 Route::get('/admin', 'admin\AdminController@dashboard');
 
-Route::get('/admin/food-categories', 'admin\FoodCategoriesController@index');
-Route::get('/admin/food-categories/create', 'admin\FoodCategoriesController@create');
-Route::post('/admin/food-categories', 'admin\FoodCategoriesController@store');
-Route::get('/admin/food-categories/{id}/edit', 'admin\FoodCategoriesController@edit');
-Route::put('/admin/food-categories/{id}/edit', 'admin\FoodCategoriesController@update');
-Route::delete('/admin/food-categories/{id}/delete', 'admin\FoodCategoriesController@delete');
+Route::get('/admin/food-categories', 'admin\FoodCategoriesController@index')->middleware('role:Admin');
+Route::get('/admin/food-categories/create', 'admin\FoodCategoriesController@create')->middleware('role:Admin');
+Route::post('/admin/food-categories', 'admin\FoodCategoriesController@store')->middleware('role:Admin');
+Route::get('/admin/food-categories/{id}/edit', 'admin\FoodCategoriesController@edit')->middleware('role:Admin');
+Route::put('/admin/food-categories/{id}/edit', 'admin\FoodCategoriesController@update')->middleware('role:Admin');
+Route::delete('/admin/food-categories/{id}/delete', 'admin\FoodCategoriesController@delete')->middleware('role:Admin');
 
-Route::get('/admin/food-items', 'admin\FoodItemsController@index');
-Route::get('/admin/food-items/create', 'admin\FoodItemsController@create');
-Route::post('/admin/food-items', 'admin\FoodItemsController@store');
-Route::get('/admin/food-items/{id}/edit', 'admin\FoodItemsController@edit');
-Route::put('/admin/food-items/{id}', 'admin\FoodItemsController@update');
-Route::delete('/admin/food-items/{id}/delete', 'admin\FoodItemsController@delete');
+Route::get('/admin/food-items', 'admin\FoodItemsController@index')->middleware('role:Admin');
+Route::get('/admin/food-items/create', 'admin\FoodItemsController@create')->middleware('role:Admin');
+Route::post('/admin/food-items', 'admin\FoodItemsController@store')->middleware('role:Admin');
+Route::get('/admin/food-items/{id}/edit', 'admin\FoodItemsController@edit')->middleware('role:Admin');
+Route::put('/admin/food-items/{id}', 'admin\FoodItemsController@update')->middleware('role:Admin');
+Route::delete('/admin/food-items/{id}/delete', 'admin\FoodItemsController@delete')->middleware('role:Admin');
 
 // Admin/Customers
 Route::get('/admin/members', 'admin\MemberController@index');
@@ -62,12 +67,20 @@ Route::delete('/admin/reservations/{id}/delete', 'admin\ReservationController@de
 
 
 // Admin/Users
-Route::get('/admin/users/', 'admin\UsersController@index');
-Route::get('/admin/users/create', 'admin\UsersController@create');
-Route::post('/admin/users/create', 'admin\UsersController@store');
-Route::get('/admin/users/{id}/edit', 'admin\UsersController@edit');
-Route::put('/admin/users/{id}/edit', 'admin\UsersController@update');
-Route::delete('/admin/users/{id}/delete', 'admin\UsersController@delete');
+Route::get('/admin/users/', 'admin\UsersController@index')->middleware('role:Admin');
+Route::get('/admin/users/create', 'admin\UsersController@create')->middleware('role:Admin');
+Route::post('/admin/users/create', 'admin\UsersController@store')->middleware('role:Admin');
+Route::get('/admin/users/{id}/edit', 'admin\UsersController@edit')->middleware('role:Admin');
+Route::put('/admin/users/{id}/edit', 'admin\UsersController@update')->middleware('role:Admin');
+Route::delete('/admin/users/{id}/delete', 'admin\UsersController@delete')->middleware('role:Admin');
+
+// Admin/Settings
+Route::get('/admin/settings/general', 'admin\SettingController@general')->middleware('role:Admin');
+Route::put('/admin/settings/general', 'admin\SettingController@updateGeneral')->middleware('role:Admin');
+Route::get('/admin/settings/seo', 'admin\SettingController@SEO')->middleware('role:Admin');
+Route::put('/admin/settings/seo', 'admin\SettingController@updateSEO')->middleware('role:Admin');
+Route::get('/admin/settings/social', 'admin\SettingController@social')->middleware('role:Admin');
+Route::put('/admin/settings/social', 'admin\SettingController@updateSocial')->middleware('role:Admin');
 
 
 
@@ -92,3 +105,17 @@ Route::get('/offers', 'StaticPagesController@about');
 Auth::routes();
 
 Route::get('/home', 'HomeController@index')->name('home');
+
+View::composer(['home', 'layouts/landing', 'pages/about', 'pages/contact', 'pages/offers', 'pages/reservations', 'pages/thank-you', 'pages/contact', 'menu/menuindex', 'menu/single-menu' ], function($view){
+    $generalSetting = GeneralSetting::find(1);
+    $socialSetting = SocialSetting::find(1);
+    $seoSetting = SeoSetting::find(1);
+
+    $view->with('settings', [
+        "general" => $generalSetting,
+        "social" => $socialSetting,
+        "seo" => $seoSetting
+    ]);
+
+
+});
